introduce enum class BinaryInputId

+++ src/handler/system_states.cpp
@@ -695,7 +695,7 @@ void SystemState::stateBinaryinput() {
   uint8_t inputIndex = (uint8_t)iiNode->getIntegerValue();
   auto&& devInput = pDev->getBinaryInput(inputIndex);
 
-  if (devInput->m_inputId != 15) {
+  if (static_cast<int>(devInput->m_inputId) != 15) { // TODO(someday): introduce enum constant
     return;
   }
